Last minut refactor
