+++ tests/htmlcov/_home_adam_PycharmProjects_Code_Challenge_fileIO_file_writer_py.html
@@ -67,31 +67,17 @@
     <table>
         <tr>
             <td class="linenos">
-<p id="n1" class="pln"><a href="#n1">1</a></p>
-<p id="n2" class="pln"><a href="#n2">2</a></p>
-<p id="n3" class="pln"><a href="#n3">3</a></p>
-<p id="n4" class="pln"><a href="#n4">4</a></p>
-<p id="n5" class="pln"><a href="#n5">5</a></p>
-<p id="n6" class="pln"><a href="#n6">6</a></p>
-<p id="n7" class="pln"><a href="#n7">7</a></p>
-<p id="n8" class="stm run hide_run"><a href="#n8">8</a></p>
-<p id="n9" class="stm run hide_run"><a href="#n9">9</a></p>
-<p id="n10" class="stm run hide_run"><a href="#n10">10</a></p>
-<p id="n11" class="stm run hide_run"><a href="#n11">11</a></p>
+<p id="n1" class="stm run hide_run"><a href="#n1">1</a></p>
+<p id="n2" class="stm run hide_run"><a href="#n2">2</a></p>
+<p id="n3" class="stm run hide_run"><a href="#n3">3</a></p>
+<p id="n4" class="stm run hide_run"><a href="#n4">4</a></p>
 
             </td>
             <td class="text">
-<p id="t1" class="pln"><span class="com"># def write_list_into_file(word_list_to_write, path_to_file):</span><span class="strut">&nbsp;</span></p>
-<p id="t2" class="pln"><span class="com">#     out_to_file = open(path_to_file, "w")</span><span class="strut">&nbsp;</span></p>
-<p id="t3" class="pln"><span class="com">#     for word in word_list_to_write:</span><span class="strut">&nbsp;</span></p>
-<p id="t4" class="pln"><span class="com">#         out_to_file.write(word + "\n")</span><span class="strut">&nbsp;</span></p>
-<p id="t5" class="pln"><span class="com">#     out_to_file.close()</span><span class="strut">&nbsp;</span></p>
-<p id="t6" class="pln"><span class="strut">&nbsp;</span></p>
-<p id="t7" class="pln"><span class="strut">&nbsp;</span></p>
-<p id="t8" class="stm run hide_run"><span class="key">def</span> <span class="nam">write_solution_into_file</span><span class="op">(</span><span class="nam">sentence</span><span class="op">,</span> <span class="nam">path_to_file</span><span class="op">)</span><span class="op">:</span><span class="strut">&nbsp;</span></p>
-<p id="t9" class="stm run hide_run">    <span class="nam">out_to_file</span> <span class="op">=</span> <span class="nam">open</span><span class="op">(</span><span class="nam">path_to_file</span><span class="op">,</span> <span class="str">"a"</span><span class="op">)</span><span class="strut">&nbsp;</span></p>
-<p id="t10" class="stm run hide_run">    <span class="nam">out_to_file</span><span class="op">.</span><span class="nam">write</span><span class="op">(</span><span class="nam">sentence</span> <span class="op">+</span> <span class="str">"\n"</span><span class="op">)</span><span class="strut">&nbsp;</span></p>
-<p id="t11" class="stm run hide_run">    <span class="nam">out_to_file</span><span class="op">.</span><span class="nam">close</span><span class="op">(</span><span class="op">)</span><span class="strut">&nbsp;</span></p>
+<p id="t1" class="stm run hide_run"><span class="key">def</span> <span class="nam">write_solution_into_file</span><span class="op">(</span><span class="nam">sentence</span><span class="op">,</span> <span class="nam">path_to_file</span><span class="op">)</span><span class="op">:</span><span class="strut">&nbsp;</span></p>
+<p id="t2" class="stm run hide_run">    <span class="nam">out_to_file</span> <span class="op">=</span> <span class="nam">open</span><span class="op">(</span><span class="nam">path_to_file</span><span class="op">,</span> <span class="str">"a"</span><span class="op">)</span><span class="strut">&nbsp;</span></p>
+<p id="t3" class="stm run hide_run">    <span class="nam">out_to_file</span><span class="op">.</span><span class="nam">write</span><span class="op">(</span><span class="nam">sentence</span> <span class="op">+</span> <span class="str">"\n"</span><span class="op">)</span><span class="strut">&nbsp;</span></p>
+<p id="t4" class="stm run hide_run">    <span class="nam">out_to_file</span><span class="op">.</span><span class="nam">close</span><span class="op">(</span><span class="op">)</span><span class="strut">&nbsp;</span></p>
 
             </td>
         </tr>
@@ -102,7 +88,7 @@
     <div class="content">
         <p>
             <a class="nav" href="index.html">&#xab; index</a> &nbsp; &nbsp; <a class="nav" href="https://coverage.readthedocs.io">coverage.py v4.2</a>,
-            created at 2016-10-24 17:29
+            created at 2016-10-24 19:49
         </p>
     </div>
 </div>
