+++ tests/htmlcov/_home_adam_PycharmProjects_Code_Challenge_preproccessing_word_prime_converter_py.html
@@ -188,7 +188,7 @@
     <div class="content">
         <p>
             <a class="nav" href="index.html">&#xab; index</a> &nbsp; &nbsp; <a class="nav" href="https://coverage.readthedocs.io">coverage.py v4.2</a>,
-            created at 2016-10-24 17:29
+            created at 2016-10-24 19:49
         </p>
     </div>
 </div>
