+++ tests/htmlcov/index.html
@@ -65,11 +65,11 @@
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>324</td>
+                <td>327</td>
                 <td>3</td>
                 <td>0</td>
                 
-                <td class="right" data-ratio="321 324">99%</td>
+                <td class="right" data-ratio="324 327">99%</td>
             </tr>
         </tfoot>
         <tbody>
@@ -175,15 +175,24 @@
             
             <tr class="file">
                 <td class="name left"><a href="_home_adam_PycharmProjects_Code_Challenge_processing_search_algoritms_py.html">/home/adam/PycharmProjects/Code_Challenge/processing/search_algoritms.py</a></td>
-                <td>83</td>
+                <td>23</td>
                 <td>0</td>
                 <td>0</td>
                 
-                <td class="right" data-ratio="83 83">100%</td>
+                <td class="right" data-ratio="23 23">100%</td>
             </tr>
             
             <tr class="file">
-                <td class="name left"><a href="_home_adam_PycharmProjects_Code_Challenge_solvers___init___py.html">/home/adam/PycharmProjects/Code_Challenge/solvers/__init__.py</a></td>
+                <td class="name left"><a href="_home_adam_PycharmProjects_Code_Challenge_processing_word_checking_py.html">/home/adam/PycharmProjects/Code_Challenge/processing/word_checking.py</a></td>
+                <td>63</td>
+                <td>0</td>
+                <td>0</td>
+                
+                <td class="right" data-ratio="63 63">100%</td>
+            </tr>
+            
+            <tr class="file">
+                <td class="name left"><a href="_home_adam_PycharmProjects_Code_Challenge_solver___init___py.html">/home/adam/PycharmProjects/Code_Challenge/solver/__init__.py</a></td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
@@ -192,7 +201,7 @@
             </tr>
             
             <tr class="file">
-                <td class="name left"><a href="_home_adam_PycharmProjects_Code_Challenge_solvers_anagram_solver_py.html">/home/adam/PycharmProjects/Code_Challenge/solvers/anagram_solver.py</a></td>
+                <td class="name left"><a href="_home_adam_PycharmProjects_Code_Challenge_solver_anagram_solver_py.html">/home/adam/PycharmProjects/Code_Challenge/solver/anagram_solver.py</a></td>
                 <td>45</td>
                 <td>3</td>
                 <td>0</td>
@@ -230,7 +239,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io">coverage.py v4.2</a>,
-            created at 2016-10-24 18:10
+            created at 2016-10-24 19:49
         </p>
     </div>
 </div>
